fix: handle unsigned 64-bit manifest IDs throughout the pipeline (#9)
add Util file path tests for save/load with manifest IDs > Long.MAX_VALUE

+++ src/main/java/in/dragonbra/javasteam/types/KeyValue.kt
@@ -580,7 +580,7 @@ class KeyValue @JvmOverloads constructor(
                     Type.POINTER,
                     -> localCurrent.value = input.readInt().toString()
 
-                    Type.UINT64 -> localCurrent.value = input.readLong().toString()
+                    Type.UINT64 -> localCurrent.value = input.readLong().toULong().toString()
                     Type.FLOAT32 -> localCurrent.value = input.readFloat().toString()
                     Type.INT64 -> localCurrent.value = input.readLong().toString()
                     else -> return false
